==> 201704TO201802CodeBackup-master/201704TO201802CodeBackup-master/201711_Tianchi_蚂蚁wifi定位第一赛季/code/5.anlays_error.py <==
import pandas as pd


anlayse = pd.read_csv('./data/offline_train.csv')
# anlayse['true'] = (anlayse['shop_id']==anlayse['shop_id_y'])

print(anlayse.groupby(['label']).describe().reset_index().to_csv('./1.csv',index=False))
